Improve annotation tooltip and editor auto positioning, reorg code
